Move createNewUsersWithMemberships to membership repo
